Avoid 'PHP Notice: Indirect modification of overloaded property'
In MW 1.37, accessing dynamic fields on Parser goes through
DeprecationHelper, which overrides __get() to return a placeholder
value if the field wasn't yet defined - causing a
'PHP Notice: Indirect modification of overloaded property' if an
attempt is made to mutate such a field without declaring it first.
Fix it by explicitly setting mExtRegexFun on first access.

Change-Id: Ieaa170b543201601abae2d6bec40235b495b66e4